requirements.txt
streamlit
google-generativeai
python-dotenv
requests
Pillow
langchain
langchain-community
faiss-cpu
tiktoken
PyPDF2
